Frontend Component Cleanup - Grouping together the user interaction parts
The Home page have been doing a bit too many features lately, so let's try to group together components that are related. One possible hing is to group together our main UX feature - the user interaction into it's separate component. This will also allow us to refactor and make those interaction bits more aligned to each other
