+++ client/src/components/PopUpAlert.js
@@ -1,4 +1,3 @@
-import * as React from "react";
 import Snackbar from "@mui/material/Snackbar";
 import MuiAlert from "@mui/material/Alert";
 
@@ -11,13 +10,10 @@ export default function PopUpAlert(props) {
 	};
 
 	return (
-		<Snackbar open={open} autoHideDuration={6000} onClose={handleClose}>
+		<Snackbar open={true} autoHideDuration={6000} onClose={handleClose}>
 			<MuiAlert
 				variant="filled"
-				elevation={6}
-				onClose={handleClose}
 				severity={props.message.success ? "success" : "error"}
-				sx={{ width: "100%" }}
 			>
 				{props.message.message}
 			</MuiAlert>
