+++ client/src/components/SubmitButton.js
@@ -1,4 +1,3 @@
-import * as React from "react";
 import Stack from "@mui/material/Stack";
 import Button from "@mui/material/Button";
 
@@ -11,7 +10,6 @@ export default function SubmitButton(props) {
 				disabled={props.selectedSuggestion == null}
 				onClick={() => {
 					props.submitButton();
-					props.handleNonAuthSubmitClick();
 				}}
 			>
 				{props.text}
