+++ client/src/pages/Home.js
@@ -1,18 +1,13 @@
 import { useEffect, useState } from "react";
-//import { Link } from "react-router-dom";
 
 import "./Home.css";
+
 import OriginalSentence from "../components/OriginalSentence";
-import SuggestionSentence from "../components/SuggestionSentence";
 import NextSentence from "../components/NextSentence";
-import NoneOfTheSuggestions from "../components/NoneOfTheSuggestions";
-import SubmitSuggestion from "../components/SubmitSuggestion";
 
 import Navbar from "../components/Navbar";
 
-import LoadingSuggestions from "../components/LoadingSuggestions";
-import UserSuggestion from "../components/UserSuggestion";
-import IsOriginalSentenceCorrect from "../components/IsOriginalSentenceCorrect";
+import UserInteraction from "../components/UserInteraction";
 import LoginDialog from "../components/LoginDialog";
 import PopUpAlert from "../components/PopUpAlert";
 
@@ -26,21 +21,17 @@ export function Home({ user }) {
 	const [alertMessage, setAlertMessage] = useState(null);
 
 	const [nextSentenceCounter, setNextSentenceCounter] = useState(0); // counter for triggering useEffect
-	const [nextSuggestionCounter, setNextSuggestionCounter] = useState(0); // counter for triggering useEffect
 	const [submitClickCounter, setSubmitClickCounter] = useState(4);
 
-	const handleNonAuthSubmitClick = () => {
+	const loadNextSentence = () => {
 		if (!user) {
 			setSubmitClickCounter(submitClickCounter - 1);
 			if (submitClickCounter < 1) {
 				setIsLoginDialogOpen(true);
 				setSubmitClickCounter(4);
-				setSelectedSuggestion(null);
 			}
 		}
-	};
 
-	const loadNextSentence = () => {
 		setSentence(null);
 		setSelectedSuggestion(null);
 		setSuggestions([]);
@@ -52,36 +43,9 @@ export function Home({ user }) {
 			const response = await fetch("/api/sentences/random");
 			const responseData = await response.json();
 			setSentence(responseData);
-			setNextSuggestionCounter((n) => n + 1);
 		};
 		obtainRandomSentence();
 	}, [nextSentenceCounter]);
-
-	useEffect(() => {
-		const getSuggestionsFromApi = async (text) => {
-			const response = await fetch(
-				`https://angocair.garg.ed.ac.uk/best/?text=${encodeURIComponent(text)}`
-			);
-			const data = await response.json();
-			setSuggestions(data.data);
-			setAlertMessage(null);
-		};
-		if (sentence) {
-			getSuggestionsFromApi(sentence.sentence);
-		}
-	}, [sentence, nextSuggestionCounter]);
-
-	const suggestionComponents = suggestions.map((text, i) => {
-		return (
-			<SuggestionSentence
-				key={i}
-				number={i + 1}
-				sentence={sentence}
-				suggestion={text}
-				setSelectedSuggestion={setSelectedSuggestion}
-			/>
-		);
-	});
 
 	return (
 		<>
@@ -104,42 +68,15 @@ export function Home({ user }) {
 					<div className="center paddingBottom">
 						<OriginalSentence sentence={sentence} />
 					</div>
-					<div className="isOriginalDiv">
-						<IsOriginalSentenceCorrect
+					<div className="userInteraction">
+						<UserInteraction
 							sentence={sentence}
 							suggestions={suggestions}
 							selectedSuggestion={selectedSuggestion}
+							setSelectedSuggestion={setSelectedSuggestion}
+							setSuggestions={setSuggestions}
 							loadNextSentence={loadNextSentence}
-						/>
-					</div>
-					<div className="paddingBottom">
-						<h3>Suggestions :</h3>
-						{suggestionComponents.length == 0 ? (
-							<LoadingSuggestions />
-						) : (
-							<div className="grid">
-								{suggestionComponents}
-								<SubmitSuggestion
-									sentence={sentence}
-									suggestions={suggestions}
-									selectedSuggestion={selectedSuggestion}
-									loadNextSentence={loadNextSentence}
-									handleNonAuthSubmitClick={handleNonAuthSubmitClick}
-									setAlertMessage={setAlertMessage}
-								/>
-							</div>
-						)}
-					</div>
-
-					<div>
-						<NoneOfTheSuggestions loadNextSentence={loadNextSentence} />
-					</div>
-					<div>
-						<UserSuggestion
-							sentence={sentence}
-							suggestions={suggestions}
-							selectedSuggestion={selectedSuggestion}
-							loadNextSentence={loadNextSentence}
+							setAlertMessage={setAlertMessage}
 						/>
 					</div>
 					{alertMessage && (
